readme: Add contact info for organizers
In case welcome, I noted that the contact info for the organizers is
present in the footer of the website, but not where it is referenced in
`README.md` in the GitHub repository.

Adding the links there as well in case people are discovering this via
GitHub, and would like to contact the organizers but weren't sure how.