index: Fix emboldened text
It seems that in the header of the homepage, the desired style for the
text that reads "No talks. No politics. Just code." was to be bold. This
is because it is surrounded by two underscores on each side, which would
be the syntax to make it bold, if it were markdown.

This removes the underscores and uses an HTML tag to achieve the effect
instead (if that is what was intended).